Add files via upload
In this function, df is the input dataframe. The function performs the following steps:

Creates a new column "inc/dec percentage" by subtracting the values in "Level 4" from those in "Level 1" and dividing the result by "Level 1".
Replaces any missing values (null values) in the dataframe with the mean of each respective column.
Replaces the abbreviations for months in column 'B' with their full names using the apply method and a lambda function.
Replaces the values in column 'E' with "From LinkedIn" for "Came_From_LinkedIn" and "Direct_traffic" for "Landed_Directly".